Add support for PKGBUILDs without leading comment
The AUR submission guidelines require that a `PKGBUILD` starts with one
or more comment lines containing contact info of maintainers and
contributors.

The `customizepkg` script relies on that requirement, ignoring the first
line of the PKGBUILD. While that works with most packages, it fails with
some because they contain `pkgname` or similar relevant declarations in
line 1.

Examples of affected packages:

- 1password [1]
- clang15 [2]
- compiler-rt15 [3]
- python-clang15 [4]
- python-setuptools-git-versioning [5]

For the first four packages, `customizepkg` ignores the patch altogether
because the first line of each of the `PKGBUILD`s contain the `pkgname`.

For `python-setuptools-git-versioning`, `customizepkg` misreads the
`pkgname` as `python-`, because the first two lines of its `PKGBUILD`
read as follows:

```PKGBUILD
_name=setuptools-git-versioning
pkgname=python-${_name}
```

Fix the regular expression in `customizepkg` that filters the `PKGBUILD`
so it now picks up the first line, too.

Fixes #27 [6].

[1]: https://aur.archlinux.org/cgit/aur.git/tree/PKGBUILD?h=1password
[2]: https://aur.archlinux.org/cgit/aur.git/tree/PKGBUILD?h=clang15
[3]: https://aur.archlinux.org/cgit/aur.git/tree/PKGBUILD?h=compiler-rt15
[4]: https://aur.archlinux.org/cgit/aur.git/tree/PKGBUILD?h=python-clang15
[5]: https://aur.archlinux.org/cgit/aur.git/tree/PKGBUILD?h=python-setuptools-git-versioning
[6]: https://github.com/ava1ar/customizepkg/issues/27