Fix Bash parameter expansion (#41)
Bash has a feature called parameter expansion. [1]
PKGBUILDs sometimes use that feature to strip
prefixes:

```
pkgname="python-foo"
_name=${pkgname#python-}
```

Because the syntax includes a hash sign (#), the
current version of customizepkg’s regex pattern
misinterprets usage of this feature as a
line comment:

```plain

[^#]*?(?= *#|\x0a)
^^^^^^               capture the part before the `#` sign
         ^^^         cut off the rest of the line
```

The problem with this pattern is that it
interprets the `#` sign as the beginning of a
line comment, even if it’s not preceded by a
whitespace, including usages of parameter
expansion. That means the pattern cuts off in
the middle of a parameter expansion, causing
customizepkg to abort with an error message:

> /usr/bin/customizepkg: eval: line 168: unexpected EOF while looking for matching `}'

Examples of affected packages:

- discover-overlay [2]
- python-llm [3]
- python-progress [4]

This problem has also been mentioned in a comment
on GitHub issue #30 [2] (however I think that the
comment is unrelated to the multiline problem.)

Fix the issue by requiring at least one whitespace
in order to detect the start of a line comment:

```plain

.*?(?=( +#)|\x0a)
^^                   capture as much as we need
      ^^^^^          cut off only if ` #` appears
                     (hash sign preceded by space)
```

[1]: https://www.gnu.org/software/bash/manual/html_node/Shell-Parameter-Expansion.html
[2]: https://github.com/ava1ar/customizepkg/issues/30#issuecomment-2573941949
[3]: https://aur.archlinux.org/cgit/aur.git/tree/PKGBUILD?h=python-llm
[4]: https://aur.archlinux.org/cgit/aur.git/tree/PKGBUILD?h=python-progress